get & create project

+++ app/scope.go
@@ -8,25 +8,25 @@ type RequestScope interface {
 	Db() *gorm.DB
 	SetDB(db *gorm.DB)
 	RequestID() string
-	SetUserID(userId uint64)
-	UserID() uint64
+	SetUserID(userId uint)
+	UserID() uint
 }
 
 type requestScope struct {
 	db        *gorm.DB
 	requestID string
-	userID    uint64
+	userID    uint
 }
 
 func (rs *requestScope) RequestID() string {
 	return rs.requestID
 }
 
-func (rs *requestScope) SetUserID(userId uint64) {
+func (rs *requestScope) SetUserID(userId uint) {
 	rs.userID = userId
 }
 
-func (rs *requestScope) UserID() uint64 {
+func (rs *requestScope) UserID() uint {
 	return rs.userID
 }
